update Exolix for BTCPay 2.3.6

+++ BTCPayServer.Plugins.Exolix/Model/SwapRequest.cs
@@ -4,7 +4,7 @@
     {
         public string CryptoFrom { get; set; }
         public string BtcAddress { get; set; }
-        public decimal BtcAmount { get; set; }
+        public float BtcAmount { get; set; }
         public string BtcPayInvoiceId { get; set; }
     }
 }
